Fix offline utilities import path
